added a makefile

+++ C00/ex00/megaphone.cpp
@@ -19,7 +19,7 @@ int	main(int argc, char **argv)
 	{	
 		for (int i = 1; i < argc; i++)
 		{	
-			for (int j = 0; j < strlen(argv[i]); j++)
+			for (int j = 0; j < (int)strlen(argv[i]); j++)
 				std::cout << (char)toupper(argv[i][j]);
 		}
 	}
